added generation of a default avatar URL to users
